+++ include/view/view.h
@@ -5,7 +5,7 @@
 #include <view/app.h>
 #include <view/awaitable.h>
 #include <view/map.h>
-
+#include <view/context.h>
 
 void view_fatal(
     const char* message,
@@ -13,6 +13,8 @@ void view_fatal(
     const char* func,
     int lineno
 );
+
+extern PyObject* ip_address;
 
 #if defined(__LINE__) && defined(__FILE__)
 #define VIEW_FATAL(msg) view_fatal(msg, __FILE__, __func__, __LINE__)