+++ src/view/__init__.py
@@ -8,9 +8,9 @@ GitHub: https://github.com/zerointensity/view.py
 try:
     import _view
 except ImportError as e:
-    raise ImportError(
-        "_view has not been built, did you forget to compile it?"
-    ) from e
+    raise ImportError("_view has not been built, did you forget to compile it?") from e
+
+from _view import Context  # re-export
 
 from . import _codec
 from .__about__ import *
@@ -21,5 +21,5 @@ from .logging import *
 from .patterns import *
 from .response import *
 from .routing import *
-from .util import *
 from .templates import *
+from .util import *