+++ src/view/util.py
@@ -50,9 +50,7 @@ def run(app_or_path: str | App) -> None:
     try:
         target = mod[split[1]]
     except KeyError:
-        raise AttributeError(
-            f'"{split[1]}" in {app_or_path} does not exist'
-        ) from None
+        raise AttributeError(f'"{split[1]}" in {app_or_path} does not exist') from None
 
     if not isinstance(target, App):
         raise MistakeError(f"{target!r} is not an instance of view.App")
@@ -123,9 +121,7 @@ def env(key: str, *, tp: type[EnvConv] = str) -> EnvConv:
         try:
             return int(value)
         except ValueError:
-            raise EnvironmentError(
-                f"{value!r} (key {key!r}) is not int-like"
-            ) from None
+            raise EnvironmentError(f"{value!r} (key {key!r}) is not int-like") from None
 
     if tp is dict:
         try: